Switch from Meson to Maven
Maven is commonly used for Java packages; Meson is not.  Also, Meson's
Java support is incomplete and had needed workarounds to add entries to
the JAR manifest.

Historically the built artifact has been called openslide.jar.  We're
already renaming it to openslide-java-$version.jar in CI, and Maven
prefers to include the version number in the JAR filename.  For
clarity and consistency, build the artifact as openslide-java-$version.jar.

Allow Maven to add its usual additional metadata to the JAR, including a
couple MANIFEST.MF properties and a complete copy of pom.xml.

Document selecting the correct JDK version, if needed.  Maven does not
necessarily run with the latest, or default, JDK on the system (e.g.
Fedora pins it to an older LTS), so it may be necessary to explicitly set
JAVA_HOME.

We don't learn that our JDK is too old until we try to compile with it.
The Maven toolchains mechanism might prevent this, but toolchains are not
autodiscovered; they must be explicitly configured by the environment,
which often does not do this.  (In GitHub Actions, actions/setup-java does
set up a toolchains file, but Fedora and Ubuntu do not.)  If we declared a
toolchain dependency, we would thus require the user to set up a
toolchains file before building; it's easier to not do this and fail at
build time if necessary.

Maven prints warnings unless pom.xml hardcodes version numbers for every
plugin mentioned there, and then it fetches and runs exactly the specified
versions, without any way to specify version ranges or to automatically
use the latest version.  Pinning to obsolete dependency versions is an
antipattern, so mitigate this design by hardcoding plugin versions and
then enabling monthly Dependabot version updates.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>